requirements.txt
allure-behave==2.15.0
allure-python-commons==2.15.0
anyio==4.11.0
attrs==25.4.0
behave==1.3.3
bitsets==0.9.1
build==1.3.0
CacheControl==0.14.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
colorama==0.4.6
concepts==0.9.2
crashtest==0.4.1
cucumber-expressions==18.0.1
cucumber-tag-expressions==8.0.0
distlib==0.4.0
dulwich==0.24.6
fastjsonschema==2.21.2
features==0.5.12
fileconfig==0.7
filelock==3.20.0
findpython==0.7.0
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
more-itertools==10.8.0
msgpack==1.1.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pbs-installer==2025.10.14
pkginfo==1.12.1.2
platformdirs==4.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
pycparser==2.23
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
RapidFuzz==3.14.1
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.37.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.3
trio==0.31.0
trio-websocket==0.12.2
trove-classifiers==2025.9.11.17
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
zstandard==0.25.0
